+++ NEWS.md
@@ -3,6 +3,7 @@
 * Added support for a minimal LCARS theme Shiny UI.
 * New functions: `lcarsPage`, `lcarsButton`, `lcarsCheckbox`, `lcarsToggle`, `buttonColumn`, `lcarsRadio`, `lcarsHeader` and other header functions, `lcarsBox`, `lcarsWell` and `lcarsApp`.
 * Added named LCARS colors from different palettes; see `lcarsdata`.
+* Added basic rectangle, pill and bracket elements: `lcarsRect`, `lcarsPill`, `lcarsLeftPill`, `lcarsRightPill`, `lcarsBracket`.
 * Added readme and vignette content.
 
 # lcars 0.1.0